add new columns to tree with live updates
This feature adds the columns 'Value','Description', and 'DataType' to
the node tree. These columns require multithreading to load their
values they need to query each node in order to find the data which is a
relatively slow operation. This feature also adds live updates to the
value column via a subscription. These subscriptions also show up in the
attributes window, so updates can be viewed live in this window as well.